global categories, store_categories table, new endpoints

+++ internal/domain/usecase/category_usecase/category_usecase.go
@@ -8,60 +8,46 @@ import (
 
 type CategoryUseCase struct {
 	categoryRepo repository.CategoryRepository
-	storeRepo    repository.StoreRepository
 	userRepo     repository.UserRepository
 }
 
-func NewCategoryUseCase(categoryRepo repository.CategoryRepository, storeRepo repository.StoreRepository, userRepo repository.UserRepository) *CategoryUseCase {
-	return &CategoryUseCase{categoryRepo: categoryRepo, storeRepo: storeRepo, userRepo: userRepo}
+func NewCategoryUseCase(
+	categoryRepo repository.CategoryRepository,
+	userRepo repository.UserRepository,
+) *CategoryUseCase {
+	return &CategoryUseCase{
+		categoryRepo: categoryRepo,
+		userRepo:     userRepo,
+	}
 }
 
 func (c *CategoryUseCase) CreateCategory(category entities.Category, uid uint64) error {
-	storeExists, err := c.storeRepo.IsExist(category.StoreID)
-	if err != nil || !storeExists {
-		return errors.New("store not found")
-	}
-
-	isOwner, err := c.userRepo.IsOwnsStore(uid, category.StoreID)
-	if err != nil || !isOwner {
-		return errors.New("user does not owning this store")
+	userData, err := c.userRepo.FindByID(uid)
+	if err != nil || !userData.IsSeller {
+		return errors.New("user is not seller")
 	}
 
 	return c.categoryRepo.Save(category)
 }
 
-func (c *CategoryUseCase) GetCategoryByID(id uint64) (entities.Category, error) {
-	return c.categoryRepo.FindByID(id)
-}
-
 func (c *CategoryUseCase) UpdateCategory(category entities.Category, uid uint64) error {
+	userData, err := c.userRepo.FindByID(uid)
+	if err != nil || !userData.IsSeller {
+		return errors.New("user is not seller")
+	}
+
 	categoryExists, err := c.categoryRepo.IsExist(category.ID)
 	if err != nil || !categoryExists {
 		return errors.New("category not found")
-	}
-
-	storeExists, err := c.storeRepo.IsExist(category.StoreID)
-	if err != nil || !storeExists {
-		return errors.New("store not found")
-	}
-
-	isOwner, err := c.userRepo.IsOwnsStore(uid, category.StoreID)
-	if err != nil || !isOwner {
-		return errors.New("user does not owning this store")
 	}
 
 	return c.categoryRepo.Update(category)
 }
 
 func (c *CategoryUseCase) DeleteCategory(id, uid uint64) error {
-	category, err := c.categoryRepo.FindByID(id)
-	if err != nil {
-		return errors.New("category not found")
-	}
-
-	isOwner, err := c.userRepo.IsOwnsStore(uid, category.StoreID)
-	if err != nil || !isOwner {
-		return errors.New("user does not owning this store")
+	userData, err := c.userRepo.FindByID(uid)
+	if err != nil || !userData.IsSeller {
+		return errors.New("user is not seller")
 	}
 
 	return c.categoryRepo.Delete(id)
